minor source/file restructuring console window focus state (the window can still be visible but input events will not be processed) optional transparency handling while mouse over or unfocused
