+++ gradle.properties
@@ -1,0 +1,2 @@
+ossrhUsername=
+ossrhPassword=
